requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
defusedxml==0.7.1
Django==5.0.1
django-allauth==0.60.1
django-appconf==1.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
django-imagekit==5.0.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.3.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
dlib==19.24.4
docopt==0.6.2
drf-yasg==1.21.7
face-recognition==1.3.0
face_recognition_models==0.3.0
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
oauthlib==3.2.2
openapi-codec==1.3.2
opencv-python==4.9.0.80
packaging==23.2
pilkit==3.0
pillow==10.2.0
pipreqs==0.4.13
postgis==1.0.4
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
yarg==0.1.9
